move translate stuff to a `features` directory
